Move Docker Cloud groovy snippet into a helper library
