+++ .rubocop.yml
@@ -29,8 +29,13 @@ Metrics/AbcSize:
   Exclude:
     - 'libraries/default.rb'
 
+Metrics/ModuleLength:
+  Max: 150
+  Exclude:
+    - 'libraries/docker.rb'
+
 Metrics/MethodLength:
-  Max: 30
+  Max: 40
   Exclude:
     - 'libraries/default.rb'
 
@@ -42,6 +47,6 @@ Style/Documentation:
   Exclude:
     - 'spec/**/*'
     - 'test/**/*'
-    - 'libraries/powerci.rb'
+    - 'libraries/docker.rb'
     - 'recipes/powerci.rb'
     - 'recipes/ibmz_ci.rb'
